final fixes for RequirementAgent and imports
