+++ .deepeval/.deepeval_telemetry.txt
@@ -1,0 +1,4 @@
+DEEPEVAL_ID=2c396aaf-38e5-49bd-8edf-0c8a81741807
+DEEPEVAL_STATUS=old
+DEEPEVAL_LAST_FEATURE=evaluation
+DEEPEVAL_EVALUATION_STATUS=old
