+++ .deepeval/.temp_test_run_data.json
@@ -1,0 +1,1 @@
+{"testCases": [{"name": "test_case_0", "input": "Who was the widow affected by the Same-sex marriage decision ruled on by the Sixth Circuit from Ohio?", "actualOutput": "The Supreme Court ruled in favor of Phelps, upholding the Fourth Circuit's decision.", "expectedOutput": "Jim Obergefell", "context": [" The case was named after plaintiff Jim Obergefell, who sued the state of Ohio after officials refused to recognize his marriage on the death certificate of his husband.", "On April 28, 2015, the Supreme Court of the United States heard oral arguments for \"Obergefell v. Hodges\" (Ohio), which was consolidated with three other same-sex marriage cases from the other states in the Sixth Circuit: \"Tanco v. Haslam\" (Tennessee), \"DeBoer v. Snyder\" (Michigan), \"Bourke v. Beshear\" (Kentucky)."], "retrievalContext": [], "toolsCalled": [{"name": "Wikipedia", "inputParameters": {}}, {"name": "Python", "inputParameters": {}}], "expectedTools": [{"name": "Wikipedia", "inputParameters": {"query": "Same-sex marriage in the Sixth Circuit"}}, {"name": "Wikipedia", "inputParameters": {"query": "Same-sex marriage in Ohio"}}], "success": false, "metricsData": [{"name": "Exact Match", "threshold": 1.0, "success": false, "score": 0.0, "reason": "The actual and expected outputs are different.", "strictMode": false}, {"name": "AnswerLLMJudgeMetric", "threshold": 0.7, "success": false, "strictMode": false, "error": "Timed out/cancelled while evaluating metric. Increase DEEPEVAL_PER_TASK_TIMEOUT_SECONDS_OVERRIDE or set DEEPEVAL_LOG_STACK_TRACES=1 for full traceback."}, {"name": "Answer Relevancy", "threshold": 0.7, "success": false, "strictMode": false, "evaluationModel": "llama3.1:8b (ollama)", "error": "Timed out/cancelled while evaluating metric. Increase DEEPEVAL_PER_TASK_TIMEOUT_SECONDS_OVERRIDE or set DEEPEVAL_LOG_STACK_TRACES=1 for full traceback."}, {"name": "Contextual Recall", "threshold": 0.7, "success": false, "strictMode": false, "evaluationModel": "llama3.1:8b (ollama)", "error": "Timed out/cancelled while evaluating metric. Increase DEEPEVAL_PER_TASK_TIMEOUT_SECONDS_OVERRIDE or set DEEPEVAL_LOG_STACK_TRACES=1 for full traceback."}, {"name": "Faithfulness", "threshold": 0.7, "success": false, "strictMode": false, "evaluationModel": "llama3.1:8b (ollama)", "error": "Timed out/cancelled while evaluating metric. Increase DEEPEVAL_PER_TASK_TIMEOUT_SECONDS_OVERRIDE or set DEEPEVAL_LOG_STACK_TRACES=1 for full traceback."}, {"name": "Tool Correctness", "threshold": 0.5, "success": true, "score": 0.5, "reason": "[\n\t Tool Calling Reason: Incomplete tool usage: missing tools [ToolCall(\n    name=\"Wikipedia\",\n    input_parameters={\n        \"query\": \"Same-sex marriage in Ohio\"\n    }\n), ToolCall(\n    name=\"Wikipedia\",\n    input_parameters={\n        \"query\": \"Same-sex marriage in the Sixth Circuit\"\n    }\n)]; expected ['Wikipedia', 'Wikipedia'], called ['Wikipedia', 'Python']. See more details above.\n\t Tool Selection Reason: No available tools were provided to assess tool selection criteria\n]\n", "strictMode": false, "verboseLogs": "Expected Tools:\n[\n    ToolCall(\n        name=\"Wikipedia\",\n        input_parameters={\n            \"query\": \"Same-sex marriage in the Sixth Circuit\"\n        }\n    ),\n    ToolCall(\n        name=\"Wikipedia\",\n        input_parameters={\n            \"query\": \"Same-sex marriage in Ohio\"\n        }\n    )\n] \n \nTools Called:\n[\n    ToolCall(\n        name=\"Wikipedia\"\n    ),\n    ToolCall(\n        name=\"Python\"\n    )\n] \n \nAvailable Tools: [] \n \nTool Selection Score: 1.0 \n \nTool Selection Reason: No available tools were provided to assess tool selection criteria"}, {"name": "ToolUsageMetric", "threshold": 0.5, "success": false, "score": 0.375, "strictMode": false}, {"name": "Argument Correctness", "threshold": 0.7, "success": false, "strictMode": false, "evaluationModel": "llama3.1:8b (ollama)", "error": "Timed out/cancelled while evaluating metric. Increase DEEPEVAL_PER_TASK_TIMEOUT_SECONDS_OVERRIDE or set DEEPEVAL_LOG_STACK_TRACES=1 for full traceback."}, {"name": "FactsSimilarityMetric", "threshold": 0.5, "success": true, "score": 0.6, "strictMode": false}], "runDuration": 60.00596510001924, "order": 0, "additionalMetadata": {"expected_facts": [" The case was named after plaintiff Jim Obergefell, who sued the state of Ohio after officials refused to recognize his marriage on the death certificate of his husband.", "On April 28, 2015, the Supreme Court of the United States heard oral arguments for \"Obergefell v. Hodges\" (Ohio), which was consolidated with three other same-sex marriage cases from the other states in the Sixth Circuit: \"Tanco v. Haslam\" (Tennessee), \"DeBoer v. Snyder\" (Michigan), \"Bourke v. Beshear\" (Kentucky)."], "tool_usage": {"Wikipedia": 7, "Python": 1}, "expected_tool_usage": {"Wikipedia": 2}, "supporting_titles": ["Same-sex marriage in the Sixth Circuit", "Same-sex marriage in Ohio"]}}], "conversationalTestCases": [], "metricsScores": [], "runDuration": 0.0}
