+++ python/eval/agents/requirement/_utils.py
@@ -1,7 +1,8 @@
 # Copyright 2025 © BeeAI a Series of LF Projects, LLC
 # SPDX-License-Identifier: Apache-2.0
 
-from typing import Any, TypeVar
+from collections import Counter
+from typing import Any, List, TypeVar
 
 from deepeval.test_case import ConversationalTestCase, LLMTestCase, ToolCall, Turn
 from pydantic import BaseModel
@@ -71,3 +72,72 @@ def to_conversation_test_case(agent: RequirementAgent, turns: list[Turn]) -> Con
             "agent_name": agent.meta.name,
         },
     )
+
+
+class EvaluationRow(BaseModel):
+    """
+    Represents a single row in the evaluation table.
+    """
+    test_case_label: str
+    # Map of metric_name -> success (bool)
+    results: dict[str, bool]
+
+
+class EvaluationTable(BaseModel):
+    """
+    Structured representation of evaluation results for consistent table reporting.
+    """
+    metric_names: List[str]
+    rows: List[EvaluationRow]
+
+
+def print_evaluation_table(table: EvaluationTable) -> None:
+    """
+    Prints the evaluation table and adds a success percentage row at the bottom.
+    """
+    metric_names = table.metric_names
+    rows = table.rows
+
+    if not rows:
+        print("No test results to display.")
+        return
+
+    total_cases = len(rows)
+    success_counts = Counter({name: 0 for name in metric_names})
+    
+    for row in rows:
+        for name in metric_names:
+            if row.results.get(name, False):
+                success_counts[name] += 1
+
+    # Footer with success percentages per metric
+    footer = ["Success %"]
+    for name in metric_names:
+        pct = (success_counts[name] / total_cases * 100) if total_cases else 0
+        footer.append(f"{pct:.0f}%")
+
+    # Header
+    header = ["Test case"] + metric_names
+    
+    # Data rows formatted for printing
+    formatted_rows = []
+    for row in rows:
+        formatted_row = [row.test_case_label]
+        for name in metric_names:
+            formatted_row.append("V" if row.results.get(name, False) else "X")
+        formatted_rows.append(formatted_row)
+
+    all_rows = [header] + formatted_rows + [footer]
+    col_widths = [max(len(str(cell)) for cell in col) for col in zip(*all_rows)]
+
+    def fmt_row(row_data):
+        return " | ".join(str(cell).ljust(w) for cell, w in zip(row_data, col_widths))
+
+    print("\n=== Evaluation Results Table ===")
+    print(fmt_row(header))
+    print("-+-".join("-" * w for w in col_widths))
+    for row_data in formatted_rows:
+        print(fmt_row(row_data))
+    print("-+-".join("-" * w for w in col_widths))
+    print(fmt_row(footer))
+    print("=== End Table ===\n")
